BUGFIX: don't hide messages caused by moving to a new level [edwards]
Solved this issue differently than [edwards]. Rather than moving
the msg() calls I changed n_level() to remove the message which
generates a --More-- prompt if there hasn't been a chance to read
the message. This is a one line change and is more universal and
isn't subject to problems with messages longer than a single line
obscuring the hero  if the new level draws the hero on line 1.